train/requirements_train.txt
# ========================================
# REQUIREMENTS PARA TREINAMENTO F5-TTS
# ========================================
# Instale com: pip install -r train/requirements_train.txt

# ========================================
# CORE DEPENDENCIES
# ========================================
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# ========================================
# F5-TTS
# ========================================
# Instalar a partir do repositório oficial
# git+https://github.com/SWivid/F5-TTS.git
# ou via PyPI (se disponível):
f5-tts>=0.1.0

# Ou instalar manualmente:
# git clone https://github.com/SWivid/F5-TTS.git
# cd F5-TTS
# pip install -e .

# ========================================
# TRAINING DEPENDENCIES
# ========================================
accelerate>=0.25.0  # Multi-GPU, mixed precision
ema-pytorch>=0.2.0  # Exponential Moving Average
wandb>=0.15.0  # Weights & Biases (opcional)
tensorboard>=2.14.0  # TensorBoard logging

# ========================================
# DATASET PREPARATION
# ========================================
# Download de YouTube
yt-dlp>=2023.10.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pyloudnorm>=0.1.1  # Loudness normalization
pydub>=0.25.1

# Datasets (Arrow format)
datasets>=2.14.0
pyarrow>=13.0.0

# ========================================
# TRANSCRIPTION
# ========================================
# Whisper para ASR
openai-whisper>=20230918

# Text preprocessing
num2words>=0.5.12  # Números para palavras
unidecode>=1.3.7  # Normalização de texto

# ========================================
# UTILITIES
# ========================================
tqdm>=4.66.0  # Progress bars
python-dotenv>=1.0.0  # .env support
safetensors>=0.4.0  # Checkpoint format
huggingface-hub>=0.19.0  # Download de modelos

# ========================================
# OPTIONAL (para otimizações)
# ========================================
# bitsandbytes>=0.41.0  # 8-bit optimizer (economiza VRAM)
# xformers>=0.0.22  # Memory-efficient attention (se disponível)

# ========================================
# DEVELOPMENT (opcional)
# ========================================
# pytest>=7.4.0  # Testes
# black>=23.10.0  # Code formatting
# flake8>=6.1.0  # Linting
